Fix. GameManager 변수 데이터형 수정 Fix. BattleManager 리치전 전용 음악 코드 추가

+++ MyLoopHero/Assets/Jung/Scripts/Managers/GameManager.cs
@@ -28,7 +28,7 @@ public class GameManager : MonoBehaviour
     private Image exp;
 
     public const float DAYCYCLE = 24f;
-    public const float LOOPLIMIT = 10f;
+    public const int LOOPLIMIT = 10;
     public float globalTime;
     public int dayCnt;
     public int loopCnt;
